requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
deprecated==1.2.18
    # via pikepdf
img2pdf==0.5.1
    # via -r requirements.in
lxml==5.3.1
    # via pikepdf
numpy==2.2.3
    # via
    #   -r requirements.in
    #   opencv-python
opencv-python==4.11.0.86
    # via -r requirements.in
packaging==24.2
    # via pikepdf
pdf2image==1.17.0
    # via -r requirements.in
pikepdf==9.5.2
    # via img2pdf
pillow==11.1.0
    # via
    #   -r requirements.in
    #   img2pdf
    #   pdf2image
    #   pikepdf
tk==0.1.0
    # via -r requirements.in
wrapt==1.17.2
    # via deprecated
